Refactorizando codigo
